Add Validation to the wallet number, integer for money , backend providers have been back , reasons with rejections in the admin dashboard

+++ Application/Dtos/Admin/UpdateUserDto.cs
@@ -14,8 +14,8 @@ namespace EgyptOnline.Dtos
         public string? ProviderType { get; set; }
         public string? Email { get; set; }
 
-        public decimal? FreeBalance { get; set; }
-        public decimal? FrozenBalance { get; set; }
+        public int? FreeBalance { get; set; }
+        public int? FrozenBalance { get; set; }
 
         /// <summary>
         /// Expected format: "yyyy-MM-ddTHH:mm:ssZ" (ISO 8601), e.g. "2026-03-11T00:00:00Z"
